refactor code -> mainly move code into proper files
