+++ frontend/javascripts/viewer/view/left_border_tabs/components/skeleton_layer_settings.tsx
@@ -26,7 +26,7 @@ import { Model } from "viewer/singletons";
 import ButtonComponent from "viewer/view/components/button_component";
 import { confirmAsync } from "../../../../dashboard/dataset/helper_components";
 import { DummyDragHandle } from "./drag_handle";
-import LayerTransformationIcon from "./layer_transformation_icon";
+import LayerTransformationIcons from "./layer_transformation_icon";
 import { LogSliderSetting } from "./log_slider_setting";
 import NumberSliderSetting from "./number_slider_setting";
 import SwitchSetting from "./switch_setting";
@@ -190,7 +190,7 @@ export default function SkeletonLayerSettings() {
             paddingRight: 1,
           }}
         >
-          <LayerTransformationIcon layer={{ category: "skeleton", name: tracingId }} />
+          <LayerTransformationIcons layer={{ category: "skeleton", name: tracingId }} />
           {!isOnlyAnnotationLayer ? (
             <ButtonComponent
               variant="text"
